consolidate AppTopBar
