+++ frontend/src/components/layout/AppTopBar/index.tsx
@@ -1,3 +1,1 @@
 export { AppTopBar } from "./AppTopBar";
-export { LogoLink } from "./LogoLink";
-export { ThemeToggleButton } from "./ThemeToggleButton";
