--- frontend/src/mitra/components/svg/Circuit.tsx
@@ -1,22 +1,0 @@
-import * as React from "react";
-
-interface Props extends React.SVGProps<SVGSVGElement> {
-  fill?: string;
-  scaler?: number;
-}
-
-const CircuitIcon = ({ fill = "white", scaler = 1, ...props }: Props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    xmlSpace="preserve"
-    width={80 * scaler}
-    height={80 * scaler}
-    fill={fill}
-    viewBox="0 0 1200 1200"
-    {...props}
-  >
-    <path d="M618.75 1053.3V900.02h-37.5v153.28a56.248 56.248 0 0 0-35.035 36.57 56.25 56.25 0 1 0 107.571 0 56.245 56.245 0 0 0-35.035-36.57zM600 1125.019a18.752 18.752 0 0 1-18.75-18.75 18.752 18.752 0 0 1 32.008-13.258 18.752 18.752 0 0 1 5.492 13.258 18.763 18.763 0 0 1-18.75 18.75zM871.88 1087.5a55.89 55.89 0 0 0-17.238 2.727l-85.887-97.973V900h-37.5v99.309c0 4.546 1.652 8.937 4.652 12.359l88.887 101.4a56.208 56.208 0 0 0-6.492 47.527 56.202 56.202 0 0 0 32.547 35.234 56.216 56.216 0 0 0 47.891-2.715 56.204 56.204 0 0 0 28.36-38.688 56.213 56.213 0 0 0-11.82-46.488 56.215 56.215 0 0 0-43.399-20.438zm0 75a18.752 18.752 0 0 1-18.75-18.75 18.752 18.752 0 0 1 32.008-13.258 18.752 18.752 0 0 1 5.492 13.258 18.763 18.763 0 0 1-18.75 18.75zM581.25 146.72V300h37.5V146.72a56.248 56.248 0 0 0 35.035-36.57 56.25 56.25 0 1 0-107.571 0 56.245 56.245 0 0 0 35.035 36.57zM600 75.001a18.752 18.752 0 0 1 18.75 18.75 18.752 18.752 0 0 1-32.008 13.258 18.752 18.752 0 0 1-5.492-13.258A18.763 18.763 0 0 1 600 75.001zM328.12 112.5a55.89 55.89 0 0 0 17.238-2.727l85.887 97.973V300h37.5v-99.309c0-4.546-1.652-8.937-4.652-12.359l-88.887-101.4a56.205 56.205 0 0 0 6.496-47.527 56.214 56.214 0 1 0-53.583 73.095zm0-75a18.752 18.752 0 0 1 18.75 18.75 18.752 18.752 0 0 1-32.008 13.258 18.752 18.752 0 0 1-5.492-13.258 18.763 18.763 0 0 1 18.75-18.75zM93.75 543.75A56.268 56.268 0 0 0 56.266 558a56.257 56.257 0 0 0-8.527 74.45 56.24 56.24 0 0 0 73.012 16.859 56.249 56.249 0 0 0 25.968-30.559H300v-37.5H146.72a56.292 56.292 0 0 0-52.969-37.5zm0 75A18.752 18.752 0 0 1 75 600a18.752 18.752 0 0 1 32.008-13.258A18.752 18.752 0 0 1 112.5 600a18.763 18.763 0 0 1-18.75 18.75zM56.25 384.38a55.869 55.869 0 0 0 30.684-9.164l101.4 88.887c3.422 3 7.813 4.652 12.359 4.652h99.309v-37.5h-92.254l-97.973-85.887a56.245 56.245 0 0 0-4.898-45.46 56.24 56.24 0 1 0-90.406 65.89 56.239 56.239 0 0 0 41.78 18.582zm0-75A18.752 18.752 0 0 1 75 328.13a18.752 18.752 0 0 1-32.008 13.258A18.752 18.752 0 0 1 37.5 328.13a18.763 18.763 0 0 1 18.75-18.75zM1106.2 656.25a56.268 56.268 0 0 0 37.484-14.25 56.257 56.257 0 0 0 8.527-74.45 56.24 56.24 0 0 0-73.012-16.859 56.249 56.249 0 0 0-25.969 30.559H899.95v37.5h153.28a56.292 56.292 0 0 0 52.969 37.5zm0-75a18.752 18.752 0 0 1 18.75 18.75 18.752 18.752 0 0 1-32.008 13.258A18.752 18.752 0 0 1 1087.45 600a18.763 18.763 0 0 1 18.75-18.75zM1143.8 815.62a55.869 55.869 0 0 0-30.684 9.164l-101.4-88.887a18.744 18.744 0 0 0-12.359-4.652h-99.309v37.5h92.254l97.973 85.887a56.245 56.245 0 0 0 4.898 45.46 56.24 56.24 0 1 0 90.406-65.89 56.239 56.239 0 0 0-41.78-18.582zm0 75a18.752 18.752 0 0 1-18.75-18.75 18.752 18.752 0 0 1 32.008-13.258 18.752 18.752 0 0 1 5.492 13.258 18.763 18.763 0 0 1-18.75 18.75zM768.75 207.75l85.887-97.973a56.413 56.413 0 0 0 50.191-7.766 56.403 56.403 0 0 0 23.605-44.973 56.412 56.412 0 0 0-72.035-55.12 56.406 56.406 0 0 0-31.61 85.02l-88.886 101.4a18.744 18.744 0 0 0-4.652 12.358v99.31h37.5zM871.87 37.5a18.752 18.752 0 0 1 18.75 18.75 18.752 18.752 0 0 1-32.008 13.258 18.752 18.752 0 0 1-5.492-13.258 18.763 18.763 0 0 1 18.75-18.75zM393.75 862.5h412.5a56.323 56.323 0 0 0 39.758-16.492A56.32 56.32 0 0 0 862.5 806.25v-412.5a56.323 56.323 0 0 0-16.492-39.758A56.32 56.32 0 0 0 806.25 337.5h-412.5a56.323 56.323 0 0 0-39.758 16.492A56.32 56.32 0 0 0 337.5 393.75v412.5a56.323 56.323 0 0 0 16.492 39.758A56.32 56.32 0 0 0 393.75 862.5zM375 393.75A18.763 18.763 0 0 1 393.75 375h412.5A18.763 18.763 0 0 1 825 393.75v412.5A18.763 18.763 0 0 1 806.25 825h-412.5A18.763 18.763 0 0 1 375 806.25z" />
-    <path d="M520.68 712.5h158.65-.004c8.797-.012 17.23-3.508 23.45-9.727s9.714-14.652 9.726-23.449v-158.65c-.012-8.796-3.508-17.23-9.727-23.449s-14.652-9.714-23.449-9.726h-158.65c-8.797.012-17.23 3.508-23.449 9.726s-9.715 14.652-9.727 23.45v158.65c.012 8.796 3.508 17.23 9.727 23.448s14.652 9.715 23.45 9.727zm4.324-187.5h150v150h-150zM656.25 300h37.5v-46.875a18.752 18.752 0 0 0-28.125-16.238 18.754 18.754 0 0 0-9.375 16.238zM543.75 300v-46.875a18.752 18.752 0 0 0-28.125-16.238 18.754 18.754 0 0 0-9.375 16.238V300zM900 431.25v37.5h99.309c4.547 0 8.938-1.652 12.359-4.652l101.4-88.887a56.264 56.264 0 0 0 50.344 5.574 56.264 56.264 0 0 0 34.426-37.148 56.262 56.262 0 0 0-54.973-71.839 56.27 56.27 0 0 0-52.636 73.567l-97.973 85.887zm243.75-121.88a18.752 18.752 0 0 1 18.75 18.75 18.752 18.752 0 0 1-32.008 13.258A18.752 18.752 0 0 1 1125 328.12a18.763 18.763 0 0 1 18.75-18.75zM965.62 525a18.752 18.752 0 0 0-18.75-18.75h-46.875v37.5h46.875A18.752 18.752 0 0 0 965.62 525zM543.75 900h-37.5v46.875a18.752 18.752 0 0 0 28.125 16.238 18.754 18.754 0 0 0 9.375-16.238zM656.25 900v46.875a18.752 18.752 0 0 0 28.125 16.238 18.754 18.754 0 0 0 9.375-16.238V900zM965.62 675a18.752 18.752 0 0 0-18.75-18.75h-46.875v37.5h46.875A18.752 18.752 0 0 0 965.62 675zM234.38 525a18.752 18.752 0 0 0 18.75 18.75h46.875v-37.5H253.13A18.752 18.752 0 0 0 234.38 525zM253.12 693.75h46.875v-37.5H253.12a18.752 18.752 0 0 0-16.238 28.125 18.754 18.754 0 0 0 16.238 9.375zM300 768.75v-37.5h-99.309a18.743 18.743 0 0 0-12.359 4.652l-101.4 88.887a56.264 56.264 0 0 0-50.344-5.574 56.264 56.264 0 0 0-34.426 37.148 56.257 56.257 0 0 0 9.379 49.773 56.277 56.277 0 0 0 45.594 22.07 56.273 56.273 0 0 0 52.636-73.57l97.973-85.887zM56.25 890.63a18.752 18.752 0 0 1-18.75-18.75 18.752 18.752 0 0 1 32.008-13.258A18.752 18.752 0 0 1 75 871.88a18.763 18.763 0 0 1-18.75 18.75zM431.25 992.25l-85.887 97.973a56.41 56.41 0 0 0-73.8 52.739 56.412 56.412 0 0 0 109.277 20.582 56.41 56.41 0 0 0-5.629-50.48l88.887-101.4c3-3.422 4.652-7.813 4.652-12.359v-99.309h-37.5zM328.13 1162.5a18.752 18.752 0 0 1-18.75-18.75 18.752 18.752 0 0 1 32.008-13.258 18.752 18.752 0 0 1 5.492 13.258 18.763 18.763 0 0 1-18.75 18.75z" />
-  </svg>
-);
-export default CircuitIcon;
