--- frontend/src/mitra/history/added.txt
@@ -1,1 +1,0 @@
-# should remain empty
