--- frontend/src/mitra/history/deleted.txt
@@ -1,41 +1,0 @@
-frontend/content/news/author-and-translators-identification-initiative/en.mdx
-frontend/content/news/buddhanexus-in-the-digital-orientalist/de.mdx
-frontend/content/news/buddhanexus-in-the-digital-orientalist/en.mdx
-frontend/content/news/buddhanexus-published/de.mdx
-frontend/content/news/buddhanexus-published/en.mdx
-frontend/content/news/buddhanexus-updates-2021/de.mdx
-frontend/content/news/buddhanexus-updates-2021/en.mdx
-frontend/content/pages/contact/de.mdx
-frontend/content/pages/contact/en.mdx
-frontend/content/pages/events/de.mdx
-frontend/content/pages/events/en.mdx
-frontend/content/pages/history/de.mdx
-frontend/content/pages/history/en.mdx
-frontend/content/pages/institutions/de.mdx
-frontend/content/pages/institutions/en.mdx
-frontend/content/pages/introduction/de.mdx
-frontend/content/pages/introduction/en.mdx
-frontend/content/pages/people/de.mdx
-frontend/content/pages/people/en.mdx
-frontend/content/pages/presentations/de.mdx
-frontend/content/pages/presentations/en.mdx
-frontend/content/pages/projects/de.mdx
-frontend/content/pages/projects/en.mdx
-frontend/content/pages/publications/de.mdx
-frontend/content/pages/publications/en.mdx
-frontend/public/assets/images/atii_logo.png
-frontend/public/assets/images/bg_chn_upscaled_bw.jpg
-frontend/public/assets/images/bg_chn_upscaled.jpeg
-frontend/public/assets/images/bg_pli_upscaled_bw.jpg
-frontend/public/assets/images/bg_pli_upscaled.jpeg
-frontend/public/assets/images/bg_skt_upscaled_bw.jpg
-frontend/public/assets/images/bg_skt_upscaled.jpeg
-frontend/public/assets/images/bg_tib_upscaled_bw.jpg
-frontend/public/assets/images/bg_tib_upscaled.jpeg
-frontend/public/assets/images/bg_welcome_upscaled_bw.jpg
-frontend/public/assets/images/bg_welcome_upscaled.jpeg
-frontend/public/assets/images/leaf.svg
-frontend/src/components/layout/Copyright.tsx
-frontend/src/pages/atii/index.tsx
-frontend/src/pages/atii/sample-page.tsx
-frontend/src/pages/atii/sample-page.tsxfrontend/content/news/author-and-translators-identification-initiative/de.mdx
